continue splitting ExpDict up
