--- src/clstepcore/ExpDict.inline.cc
@@ -1,26 +1,0 @@
-
-/*
-* NIST STEP Core Class Library
-* clstepcore/ExpDict.inline.cc
-* April 1997
-* K. C. Morris
-* David Sauder
-
-* Development of this software was funded by the United States Government,
-* and is not subject to copyright.
-*/
-
-#include <ExpDict.h>
-#include "sc_memmgr.h"
-
-
-
-
-
-
-Used_item::~Used_item() {
-}
-
-Referenced_item::~Referenced_item() {
-}
-
